paper2_apacc_validation/Code/requirements.txt
# Core dependencies
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# CARLA Python API
carla==0.9.15

# SUMO
eclipse-sumo==1.18.0
traci==1.18.0
sumolib==1.18.0

# MATLAB Engine (optional - install separately if using MATLAB)
# matlabengine==23.2.1

# Data processing
h5py==3.9.0
netCDF4==1.6.4
xarray==2023.7.0

# Time series database
influxdb-client==1.37.0

# Task queue
redis==4.6.0
celery==5.3.1

# Monitoring
prometheus-client==0.17.1

# Logging
loguru==0.7.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Parallel processing
joblib==1.3.1
ray==2.6.1

# Statistical analysis
statsmodels==0.14.0
scikit-learn==1.3.0

# Optimization
cvxpy==1.3.2
mosek==10.0.46  # Optional commercial solver

# Visualization
plotly==5.15.0
dash==2.11.1

# Documentation
sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Code quality
black==23.7.0
flake8==6.1.0
mypy==1.4.1

# Performance profiling
py-spy==0.3.14
memory-profiler==0.61.0